+++ boards/arm/npcm400f_evb/npcm400f_evb_defconfig
@@ -88,6 +88,12 @@ CONFIG_ADC_NPCM4XX=y
 CONFIG_SPI=y
 CONFIG_SPIP_NPCM4XX=y
 
+# I3C driver
+CONFIG_I3C=y
+CONFIG_I3C_NPCM4XX=y
+CONFIG_I3C_SLAVE=y
+CONFIG_I3C_SLAVE_MQUEUE=y
+
 # USB Driver
 CONFIG_USB=y
 CONFIG_USB_DC_NPCM4XX=y
