+++ drivers/i3c/npcm4xx/i3c_core.h
@@ -5,7 +5,7 @@
 #ifndef __I3C_CORE_H__
 #define __I3C_CORE_H__
 
-#include <drivers/i3c/NPCM4XX/pub_I3C.h>
+#include "pub_I3C.h"
 
 /*
  * variable
